subgraph codegen refactor
